+++ examples/merkle/logicsigVerifier/main.go
@@ -26,6 +26,7 @@ import (
 	"github.com/giuliop/algoplonk/setup"
 	"github.com/giuliop/algoplonk/testutils"
 	sdk "github.com/giuliop/algoplonk/testutils/algosdkwrapper"
+	"github.com/giuliop/algoplonk/utils"
 	"github.com/giuliop/algoplonk/verifier"
 )
 
@@ -124,11 +125,11 @@ func main() {
 	}
 
 	fmt.Print("\nCompiling verifier logicsig with puyapy: ")
-	err = testutils.CompileWithPuyaPy(puyaVerifierFilename, "")
+	err = utils.CompileWithPuyaPy(puyaVerifierFilename, "")
 	if err != nil {
 		log.Fatal(err)
 	}
-	err = testutils.RenamePuyaPyOutput(verifier.DefaultFileName, verifierName,
+	err = utils.RenamePuyaPyOutput(verifier.DefaultFileName, verifierName,
 		artefactsFolder)
 	if err != nil {
 		log.Fatal(err)